Tehtävä 13.24, complete, naming convention back to normal. Also fixed a crash on server restart when no tables are present (clearing Sessions)

+++ index.js
@@ -12,6 +12,7 @@ const logoutRouter = require('./src/controllers/logout')
 const authorRouter = require('./src/controllers/authors')
 const readlistRouter = require('./src/controllers/readList')
 const { Session } = require('./src/models')
+const { sequelize } = require('./utils/database')
 
 app.use(express.json())
 app.use('/api/blogs', blogRouter)
@@ -21,11 +22,29 @@ app.use('/api/logout', logoutRouter)
 app.use('/api/authors', authorRouter)
 app.use('/api/readinglists', readlistRouter)
 
+const clearSessionsOnRestart = async () => {
+  try {
+    const tableExists = await sequelize.getQueryInterface().showAllSchemas()
+      .then((tableList) => {
+        return tableList.some((table) => table.tableName === 'Sessions')
+      })
+
+    if (tableExists) {
+      await Session.destroy({
+        where: {},
+        truncate: true
+      })
+      console.log('All sessions cleared')
+    } else {
+      console.log('Session table does not exist, skipping session clearing')
+    }
+  } catch (error) {
+    console.error('Error clearing sessions on server restart:', error)
+  }
+}
+
 const start = async () => {
-  await Session.destroy({
-    where: {},
-    truncate: true
-  })
+  clearSessionsOnRestart()
   await connectToDatabase()
   app.listen(PORT, () => {
     console.log(`Server running on port ${PORT}`)
